# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  domain.Document:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      document_key:
        type: string
      id:
        type: string
      project_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  http.DocumentDetailResponse:
    properties:
      document:
        $ref: '#/definitions/domain.Document'
      revisions:
        items:
          $ref: '#/definitions/domain.Revision'
        type: array
    type: object
...
  http.RetryResponse:
    properties:
      status:
        example: queued
        type: string
    type: object
  http.UpdateRequest:
    properties:
      description:
        type: string
      title:
        maxLength: 255
        type: string
      version:
        minimum: 1
        type: integer
    required:
    - title
    - version
    type: object
...
  http.UploadResponse:
    properties:
      document:
        $ref: '#/definitions/domain.Document'
      revision:
        $ref: '#/definitions/domain.Revision'
    type: object
...
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      retryable:
        type: boolean
    type: object
...
  usecase.PresignResult:
    properties:
      expires_at:
        type: string
      object_key:
        type: string
      upload_id:
        type: string
      upload_url:
        type: string
    type: object
info:
  contact: {}
  description: Boilerplate Go Clean Architecture cho ISC.
  title: docs-hub-api
  version: "1.0"
paths:
  /internal/api/v1/projects/{project_id}/documents:
    get:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - default: 1
        description: Trang
        in: query
        name: page
        type: integer
      - default: 20
        description: Số bản ghi/trang
        in: query
        name: limit
        type: integer
      - description: Tìm theo tiêu đề
        in: query
        name: q
        type: string
      - description: Trạng thái revision
        in: query
        name: status
        type: string
      - description: MIME type
        in: query
        name: type
        type: string
      - description: Project version ID
        format: uuid
        in: query
        name: version_id
        type: string
      - description: Change request ID
        format: uuid
        in: query
        name: change_request_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Document'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Danh sách tài liệu trong project
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/{id}:
    delete:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Xóa mềm tài liệu
      tags:
      - documents
    get:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/http.DocumentDetailResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Chi tiết tài liệu và các revision
      tags:
      - documents
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Metadata và version hiện tại
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/domain.Document'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Cập nhật metadata tài liệu
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/{id}/revisions:
    post:
      consumes:
      - multipart/form-data
      description: Upload multipart trực tiếp vào document có sẵn. Phải truyền đúng
        một version hoặc change request.
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: File tài liệu
        in: formData
        name: file
        required: true
        type: file
      - description: Project version ID
        format: uuid
        in: formData
        name: project_version_id
        type: string
      - description: Change request ID
        format: uuid
        in: formData
        name: change_request_id
        type: string
      - description: Kích thước khai báo
        in: formData
        name: size_bytes
        type: integer
      - description: SHA-256 dạng hex, 64 ký tự
        in: formData
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/http.UploadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Upload revision mới cho document
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/{id}/revisions/{revision_id}/download:
    get:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Revision ID
        format: uuid
        in: path
        name: revision_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Tải file revision gốc
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/{id}/revisions/{revision_id}/retry:
    post:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Revision ID
        format: uuid
        in: path
        name: revision_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/http.RetryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Chạy lại ingestion cho revision thất bại
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/{id}/revisions/{revision_id}/status:
    get:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Revision ID
        format: uuid
        in: path
        name: revision_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/domain.Revision'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Xem trạng thái ingestion của revision
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/{id}/revisions/{revision_id}/view:
    get:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Document ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Revision ID
        format: uuid
        in: path
        name: revision_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Xem file revision inline
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/uploads:
    post:
      consumes:
      - multipart/form-data
      description: Upload multipart trực tiếp. Phải truyền đúng một project_version_id
        hoặc change_request_id.
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: File TXT, MD, CSV, PDF text-layer, DOCX hoặc XLSX
        in: formData
        name: file
        required: true
        type: file
      - description: Tiêu đề document mới
        in: formData
        name: title
        required: true
        type: string
      - description: Mô tả
        in: formData
        name: description
        type: string
      - description: Document ID nếu thêm revision
        format: uuid
        in: formData
        name: document_id
        type: string
      - description: Project version ID
        format: uuid
        in: formData
        name: project_version_id
        type: string
      - description: Change request ID
        format: uuid
        in: formData
        name: change_request_id
        type: string
      - description: Kích thước khai báo để đối chiếu
        in: formData
        name: size_bytes
        type: integer
      - description: SHA-256 dạng hex, 64 ký tự
        in: formData
        name: sha256
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/http.UploadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Upload tài liệu mới
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/uploads/{upload_id}/complete:
    post:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Upload session ID
        format: uuid
        in: path
        name: upload_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/http.UploadResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Hoàn tất presigned upload
      tags:
      - documents
  /internal/api/v1/projects/{project_id}/documents/uploads/presign:
    post:
      consumes:
      - application/json
      description: Chỉ khả dụng khi storage.driver=minio; local filesystem trả HTTP
        400 và dùng multipart upload.
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: Thông tin file và scope
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.PresignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/usecase.PresignResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Tạo presigned upload
      tags:
      - documents
...
securityDefinitions:
  BearerAuth:
    description: 'Nhập: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"